requirements.txt
discord.py>=2.4.0
python-dotenv==1.0.0
cryptography==41.0.7
psutil>=5.9.0 